(WIP) run backend trigger validation before saving
