+++ src/Containers/TriggerDuplicateContainer.tsx
@@ -4,13 +4,15 @@ import { ValidationContainer } from "@skbkontur/react-ui-validations";
 import { Button } from "@skbkontur/react-ui/components/Button";
 import MoiraApi from "../Api/MoiraApi";
 import { withMoiraApi } from "../Api/MoiraApiInjection";
-import { Trigger } from "../Domain/Trigger";
+import { Trigger, ValidateTriggerResult } from "../Domain/Trigger";
 import { getPageLink } from "../Domain/Global";
 import { Config } from "../Domain/Config";
 import RouterLink from "../Components/RouterLink/RouterLink";
 import Layout, { LayoutContent, LayoutTitle } from "../Components/Layout/Layout";
 import TriggerEditForm from "../Components/TriggerEditForm/TriggerEditForm";
 import { ColumnStack, RowStack, Fit } from "../Components/ItemsStack/ItemsStack";
+import { Toast } from "@skbkontur/react-ui/components/Toast/Toast";
+import has from "lodash/has";
 
 // TODO check id wasn't undefined
 type Props = RouteComponentProps<{ id?: string }> & { moiraApi: MoiraApi };
@@ -20,7 +22,27 @@ type State = {
     trigger?: Partial<Trigger>;
     tags?: string[];
     config?: Config;
+    validationResult?: ValidateTriggerResult;
 };
+
+function getAsyncValidator() {
+    let storage;
+    return async (
+        condition: Promise<ValidateTriggerResult>,
+        callback: (value: ValidateTriggerResult) => void
+    ) => {
+        storage = condition;
+        try {
+            const result = await condition;
+
+            if (storage === condition) {
+                callback(result);
+            }
+        } catch (e) {
+            Toast.push(e.toString());
+        }
+    };
+}
 
 class TriggerDuplicateContainer extends React.Component<Props, State> {
     public state: State = {
@@ -28,6 +50,7 @@ class TriggerDuplicateContainer extends React.Component<Props, State> {
     };
 
     private validationContainer = React.createRef<ValidationContainer>();
+    private asyncValidator = getAsyncValidator();
 
     public componentDidMount() {
         document.title = "Moira - Duplicate trigger";
@@ -56,8 +79,8 @@ class TriggerDuplicateContainer extends React.Component<Props, State> {
     }
 
     render(): React.ReactElement {
-        const { match, moiraApi } = this.props;
-        const { loading, error, trigger, tags, config } = this.state;
+        const { match } = this.props;
+        const { loading, error, trigger, tags, config, validationResult } = this.state;
         return (
             <Layout loading={loading} error={error}>
                 <LayoutContent>
@@ -71,9 +94,9 @@ class TriggerDuplicateContainer extends React.Component<Props, State> {
                                             <TriggerEditForm
                                                 data={trigger}
                                                 tags={tags || []}
-                                                remoteAllowed={config.remoteAllowed}
+                                                remoteAllowed={true}
                                                 onChange={this.handleChange}
-                                                validateTrigger={moiraApi.validateTrigger}
+                                                validationResult={validationResult}
                                             />
                                         )}
                                     </ValidationContainer>
@@ -115,27 +138,32 @@ class TriggerDuplicateContainer extends React.Component<Props, State> {
     async handleSubmit() {
         let { trigger } = this.state;
         const { moiraApi, history } = this.props;
-        const isValid = await this.validateForm();
+        const isValid = (await this.validateForm()) && !has(this.state.validationResult, "targets");
         if (isValid && trigger) {
-            this.setState({ loading: true });
-            if (trigger.trigger_type === "expression") {
-                trigger = {
-                    ...trigger,
-                    error_value: null,
-                    warn_value: null,
-                };
-            }
-            if (trigger.trigger_type === "rising" || trigger.trigger_type === "falling") {
-                trigger = {
-                    ...trigger,
-                    expression: "",
-                };
-            }
-            try {
-                const { id } = await moiraApi.addTrigger(trigger);
-                history.push(getPageLink("trigger", id));
-            } catch (error) {
-                this.setState({ error: error.message, loading: false });
+            await this.handleValidateTrigger(trigger);
+            const isBackendValid = !has(this.state.validationResult, "targets");
+
+            if (isBackendValid) {
+                this.setState({ loading: true });
+                if (trigger.trigger_type === "expression") {
+                    trigger = {
+                        ...trigger,
+                        error_value: null,
+                        warn_value: null,
+                    };
+                }
+                if (trigger.trigger_type === "rising" || trigger.trigger_type === "falling") {
+                    trigger = {
+                        ...trigger,
+                        expression: "",
+                    };
+                }
+                try {
+                    const { id } = await moiraApi.addTrigger(trigger);
+                    history.push(getPageLink("trigger", id));
+                } catch (error) {
+                    this.setState({ error: error.message, loading: false });
+                }
             }
         }
     }
@@ -144,6 +172,17 @@ class TriggerDuplicateContainer extends React.Component<Props, State> {
         this.setState(
             (prevState: State) => ({ trigger: { ...prevState.trigger, ...update } }),
             callback
+        );
+    };
+
+    private handleValidateTrigger = async (trigger: Partial<Trigger>) => {
+        const { moiraApi } = this.props;
+
+        await this.asyncValidator(
+            moiraApi.validateTrigger(trigger),
+            (validationResult: ValidateTriggerResult) => {
+                this.setState({ validationResult });
+            }
         );
     };
